move things into utils file
